# specs/002-orders-payments-stabilization/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: To7fa Orders & Payments API
  description: API contract for order lifecycle, payments, and inventory management
  version: 2.0.0
  contact:
    name: To7fa Backend Team
servers:
  - url: /api/v1
    description: API v1
...
paths:
  /orders/create/:
    post:
      tags: [Orders]
      summary: Create a new order
      description: |
        Creates an order with stock reservation. 
        - For wallet/online payments: Initial state is `pending_payment`
        - For COD: Initial state is `cod_pending`
        - Idempotent: Duplicate idempotency_key returns existing order
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '200':
          description: Idempotent response - existing order returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Validation error (insufficient stock, invalid data)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
        '403':
          description: User locked or unverified
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OrderCreateRequest:
      type: object
      required:
        - items_data
        - shipping_address
        - payment_method
      properties:
        items_data:
          type: array
          minItems: 1
          items:
            type: object
            required:
              - product_id
              - quantity
            properties:
              product_id:
                type: integer
              quantity:
                type: integer
                minimum: 1
              variant_id:
                type: integer
                nullable: true
        shipping_address:
          type: string
        shipping_cost:
          type: number
          format: decimal
          default: 0
        payment_method:
          type: string
          enum: [wallet, cod, instapay, credit_card]
        idempotency_key:
          type: string
          maxLength: 255
          description: Unique key to prevent duplicate orders
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum:
            - pending_payment
            - cod_pending
            - paid
            - processing
            - shipped
            - delivered
            - completed
            - cancelled
            - refunded
            - failed
        total_amount:
          type: number
          format: decimal
        shipping_address:
          type: string
        shipping_cost:
          type: number
          format: decimal
        payment_method:
          type: string
        payment_status:
          type: boolean
        payment_timeout_at:
          type: string
          format: date-time
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItemResponse:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: decimal
        seller_id:
          type: integer
        seller_name:
          type: string
        variant_id:
          type: integer
          nullable: true
        item_status:
          type: string
          enum: [pending, processing, shipped, delivered, completed, cancelled]
        reservation_status:
          type: string
          enum: [reserved, released, committed]
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error context